Edited style on level select page.
I hate having to 'style' things in js if you can even call it that. Like the values seem pointless. I can't give it an absolute or rel measure. no. It just wants like 5. And the difference between 5 and 6 is the whole screen, but a .1 is nothing. I hate styling in JS. all hail css. Never would've thought I would have said that. But I made my bed and here we are. Love, bobby.